Fix extracting linked property for single target ref
