+++ adapters/handlers/grpc/prepare_reply.go
@@ -243,47 +243,3 @@ func getAllNonRefNonBlobProperties(scheme schema.Schema, className string) ([]se
 
 	return props, nil
 }
-
-func extractPropertiesRequest(reqProps *pb.Properties) []search.SelectProperty {
-	var props []search.SelectProperty
-	if reqProps == nil {
-		return props
-	}
-	if reqProps.NonRefProperties != nil && len(reqProps.NonRefProperties) > 0 {
-		for _, prop := range reqProps.NonRefProperties {
-			props = append(props, search.SelectProperty{
-				Name:        prop,
-				IsPrimitive: true,
-			})
-		}
-	}
-
-	if reqProps.RefProperties != nil && len(reqProps.RefProperties) > 0 {
-		for _, prop := range reqProps.RefProperties {
-			props = append(props, search.SelectProperty{
-				Name:        prop.ReferenceProperty,
-				IsPrimitive: false,
-				Refs: []search.SelectClass{{
-					ClassName:            prop.WhichCollection,
-					RefProperties:        extractPropertiesRequest(prop.LinkedProperties),
-					AdditionalProperties: extractAdditionalPropsForRefs(prop.Metadata),
-				}},
-			})
-		}
-	}
-
-	return props
-}
-
-func extractAdditionalPropsForRefs(prop *pb.AdditionalProperties) additional.Properties {
-	return additional.Properties{
-		Vector:             prop.Vector,
-		Certainty:          prop.Certainty,
-		ID:                 prop.Uuid,
-		CreationTimeUnix:   prop.CreationTimeUnix,
-		LastUpdateTimeUnix: prop.LastUpdateTimeUnix,
-		Distance:           prop.Distance,
-		Score:              prop.Score,
-		ExplainScore:       prop.ExplainScore,
-	}
-}
